+++ src/DabModulator.cpp
@@ -477,7 +477,7 @@ const string DabModulator::get_parameter(const string& parameter) const
 const json::map_t DabModulator::get_all_values() const
 {
     json::map_t map;
-    map["rate"].v = m_settings.outputRate;
-    map["num_clipped_samples"].v = m_formatConverter ? m_formatConverter->get_num_clipped_samples() : 0;
+    map["rate"].v = (uint64_t)m_settings.outputRate;
+    map["num_clipped_samples"].v = (uint64_t)(m_formatConverter ? m_formatConverter->get_num_clipped_samples() : (size_t)0);
     return map;
 }
