+++ src/output/Dexter.cpp
@@ -435,10 +435,10 @@ SDRDevice::run_statistics_t Dexter::get_run_statistics(void) const
     run_statistics_t rs;
     {
         std::unique_lock<std::mutex> lock(m_attr_thread_mutex);
-        rs["underruns"].v = underflows;
+        rs["underruns"].v = (uint64_t)underflows;
     }
-    rs["latepackets"].v = num_late;
-    rs["frames"].v = num_frames_modulated;
+    rs["latepackets"].v = (uint64_t)num_late;
+    rs["frames"].v = (uint64_t)num_frames_modulated;
 
     rs["in_holdover_since"].v = 0;
     rs["remaining_holdover_s"].v = m_conf.maxGPSHoldoverTime;
@@ -449,12 +449,12 @@ SDRDevice::run_statistics_t Dexter::get_run_statistics(void) const
             rs["clock_state"].v = "normal"; break;
         case DexterClockState::Holdover:
             rs["clock_state"].v = "holdover";
-            rs["in_holdover_since"].v = m_holdover_since_t;
+            rs["in_holdover_since"].v = (int64_t)m_holdover_since_t;
             {
                 using namespace std::chrono;
                 const auto max_holdover_duration = seconds(m_conf.maxGPSHoldoverTime);
                 const duration<double> remaining = max_holdover_duration - (steady_clock::now() - m_holdover_since);
-                rs["remaining_holdover_s"].v = (ssize_t)duration_cast<seconds>(remaining).count();
+                rs["remaining_holdover_s"].v = (int64_t)duration_cast<seconds>(remaining).count();
             }
             break;
     }
